requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==23.0.0
Flask-Login~=0.6.3
psycopg2-binary~=2.9.10
bcrypt~=4.2.1
email_validator == 2.2.0
flask_wtf ~=1.2.2

WTForms~=3.2.1